Revert "useAuthに処理を移行 と ログインとサインインのページを変更"
