--- app/(sign)/login/page.tsx
@@ -1,34 +1,0 @@
-import Link from "next/link";
-import Form from '../form';
-
-export default function Login({
-  searchParams,
-}: {
-  searchParams: { message: string };
-}) {
-  return (
-    <div className="flex flex-col justify-center flex-1 w-full gap-2 px-8 sm:max-w-md">
-      <Link
-        href="/"
-        className="absolute flex items-center px-4 py-2 text-sm no-underline rounded-md left-8 top-8 text-main bg-btn-base-10 hover:bg-btn-base-20-hover group"
-      >
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          width="24"
-          height="24"
-          viewBox="0 0 24 24"
-          fill="none"
-          stroke="currentColor"
-          strokeWidth="2"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          className="w-4 h-4 mr-2 transition-transform group-hover:-translate-x-1"
-        >
-          <polyline points="15 18 9 12 15 6" />
-        </svg>{" "}
-        Back
-      </Link>
-      <Form searchParams={searchParams} formType="sign-in" />
-    </div>
-  );
-}
