+++ src/hooks/useAuth.ts
@@ -1,54 +1,1 @@
-import { createClient } from "@/utils/supabase/server";
-import { redirect } from "next/navigation";
-import { headers } from "next/headers";
-
-
-export default function useAuth() {
-  const signIn = async (formData: FormData) => {
-    "use server";
-
-    const email = formData.get("email") as string;
-    const password = formData.get("password") as string;
-    const supabase = createClient();
-
-    const { error } = await supabase.auth.signInWithPassword({
-      email,
-      password,
-    });
-
-    if (error) {
-      throw new Error("Could not authenticate user");
-    }
-
-    setTimeout(() => {
-      redirect("/dashboard");
-    }, 1000);
-  };
-
-  const signUp = async (formData: FormData) => {
-    "use server";
-
-    const origin = headers().get("origin");
-    const email = formData.get("email") as string;
-    const password = formData.get("password") as string;
-    const supabase = createClient();
-
-    const { error } = await supabase.auth.signUp({
-      email,
-      password,
-      options: {
-        emailRedirectTo: `${origin}/auth/callback`,
-      },
-    });
-
-    if (error) {
-      return { error: "Could not authenticate user" };
-    }
-
-    setTimeout(() => {
-      redirect("/confirm-email");
-    } , 1000);
-  };
-
-  return { signIn, signUp };
-}
+//TODO: /app/login/page.ts の認証処理に関してはここで実装する
